+++ src/builtins/env.c
@@ -1,0 +1,52 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   env.c                                              :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: hstephan <hstephan@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/08/28 17:35:59 by hstephan          #+#    #+#             */
+/*   Updated: 2023/10/02 15:42:38 by hstephan         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../../include/minishell.h"
+
+int	ft_env(char **env)
+{
+	unsigned int	i;
+
+	i = 0;
+	printf("debut de env\n");
+	if (!env)
+		return (1);
+	while (env[i] != NULL)
+	{
+		printf("%s \n", env[i]);
+		i++;
+	}
+	return (0);
+}
+
+int	ft_ordonned_env(char **env)
+{
+	int		i;
+	char	*tmp;
+
+	i = 0;
+	if (!env)
+		return (-1);
+	while (env[i + 1] != NULL)
+	{
+		if (ft_strcmp(env[i], env[i + 1]) > 0)
+		{
+			tmp = env[i];
+			env[i] = env[i + 1];
+			env[i + 1] = tmp;
+			i = -1;
+		}
+		i++;
+	}
+	ft_env(env);
+	return (0);
+}
