+++ src/builtins/export.c
@@ -6,7 +6,7 @@
 /*   By: hstephan <hstephan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/28 17:35:59 by hstephan          #+#    #+#             */
-/*   Updated: 2023/10/02 11:28:48 by hstephan         ###   ########.fr       */
+/*   Updated: 2023/10/02 15:26:49 by hstephan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,8 +16,6 @@ static int	ft_parse_var(char *str, unsigned int *i)
 {
 	if (!str || str[0] == '\0')
 		return (0);
-	if (open_quote(str) == 1)
-		return (dquote());
 	if (str[0] == '=')
 		return (print_and_return("zsh: bad assignment\n", -1));
 	while (str[*i] != '\0' && str[*i] != '=')
@@ -54,28 +52,22 @@ static int	ft_change_val(char **env,
 		i++;
 	newval = ft_strndup(val, i);
 	if (!newval)
-	{
-		free(newvar);
-		return (-1);
-	}
+		return (free(newvar), -1);
 	new = ft_strjoin(newvar, newval);
-	// printf("apres join new = %s\n", new);
 	free(newvar);
 	free(newval);
 	if (!new)
 		return (-1);
 	free (env[posi]);
 	env[posi] = new;
-	return (0);
+	return (1);
 }
 
-static int	ft_create_var(char ***env, char *var)
+static int	ft_create_var(char ***env, char *var, unsigned int i)
 {
 	unsigned int	size;
 	char			**newenv;
-	unsigned int	i;
 
-	i = 0;
 	size = size_env(*env);
 	size++;
 	newenv = malloc(sizeof(char *) * (size + 1));
@@ -86,8 +78,6 @@ static int	ft_create_var(char ***env, char *var)
 		newenv[i] = (*env)[i];
 		i++;
 	}
-	// printf("avant de changer (*env)[size - 2] = %s\n)", (*env)[size - 2]);
-	// printf("avant de changer (*env)[size - 1] = %s\n)", (*env)[size - 1]);
 	newenv[size - 2] = ft_strdup(var);
 	if (!(newenv[size - 2]))
 	{
@@ -116,29 +106,23 @@ int	ft_export_one_by_one(char ***env, char *str)
 	posi = find_var(*env, var);
 	if (posi == -1)
 	{
-		// printf("avant create var, env = \n");
-		// ft_env(env);
-		if (ft_create_var(env, var) == -1)
+		if (ft_create_var(env, var, 0) == -1)
 			return (1);
-		// printf("apres create var, env = \n");
-		// ft_env(*env);
 		posi = find_var(*env, var);
 	}
 	if (ft_change_val(*env, posi, &(str[i]), i) == -1)
 		return (1);
-	// printf("apres change val, env = \n");
-	// ft_env(*env);
 	return (0);
 }
 
-int		ft_export(char ***env, char *str)
+int	ft_export(char ***env, char *str)
 {
 	char			**tab;
 	unsigned int	i;
 
 	i = 0;
 	if (!str || str[0] == '\0')
-		return(ft_ordonned_env(*env));
+		return (ft_ordonned_env(*env));
 	tab = ft_nsplit(str, ' ', '\t');
 	if (!tab)
 		return (1);
