+++ src/parsing/token.c
@@ -6,7 +6,7 @@
 /*   By: hstephan <hstephan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/20 14:31:19 by zlemery           #+#    #+#             */
-/*   Updated: 2023/09/27 11:28:16 by hstephan         ###   ########.fr       */
+/*   Updated: 2023/10/02 15:38:59 by hstephan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -97,7 +97,7 @@ int	check_redirections(char **tab)
 
 char	**init_start_cmd(t_shell *shell, char *cmd_line, int index)
 {
-	int			i;
+	int		i;
 	char	**tab;
 
 	i = -1;
