+++ src/pipex/here_doc.c
@@ -1,42 +1,23 @@
-/* ******************************************************************************* */
-/*                                                                                  */
-/*                                                        :::      ::::::::         */
-/*   here_doc.c                                         :+:      :+:    :+:         */
-/*                                                    +:+ +:+         +:+      		*/
-/*   By: zlemery <zlemery@student.42.fr>                  +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           		*/
-/*   Created: 2023/09/11 13:22:03 by zlemery              #+#    #+#             	*/
-/*   Updated: 2023/09/11 13:28:05 by zlemery             ###   ########.fr       	*/
-/*                                                                            		*/
-/* ******************************************************************************** */
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   here_doc.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: hstephan <hstephan@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/10/02 15:51:50 by hstephan          #+#    #+#             */
+/*   Updated: 2023/10/02 15:54:05 by hstephan         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
 #include "../../include/minishell.h"
-
-int	nb_heredoc(char *line)
-{
-	int	i;
-	int	count;
-
-	i = -1;
-	count = 0;
-	while (line[++i])
-	{
-		if ((line[i] == '<' && line[i - 1] != '\\')
-			&& line[i + 1] == '<')
-		{
-			count++;
-			i++;
-		}
-	}
-	return (count);
-}
 
 int	create_here(t_shell *shell, char *av)
 {
 	if (shell->nb_here == 0)
 		return (0);
 	shell->here = malloc(sizeof(t_here) * shell->nb_here);
-	if(!shell->here)
+	if (!shell->here)
 	{
 		free_shell(shell, av);
 		return (-1);
@@ -44,60 +25,11 @@ int	create_here(t_shell *shell, char *av)
 	return (0);
 }
 
-int	recup_delim2(t_shell *shell, char **tmp, int j)
-{
-	int	i;
-
-	i = 0;
-	while (tmp[i])
-	{
-		printf("%s\n", tmp[i]);
-		if (!ft_strcmp(tmp[i], "<<"))
-		{
-			shell->here[j].lim = delete_quote(tmp[i + 1]);
-			free(tmp[i]);
-			i++;
-			pipe(shell->here[j].here_pipe);
-			j++;
-		}
-		else
-			free(tmp[i]);
-		i++;
-	}
-	free(tmp);
-	return (j);
-}
-
-int	recup_delim1(t_shell *shell)
-{
-	int		i;
-	int		j;
-	char	**tmp;
-
-	i = -1;
-	j = 0;
-	while (shell->token[++i])
-	{
-		tmp = ft_split(shell->token[i], ' ');
-		if (!tmp)
-			return (-1);
-		fix_quote((signed char **)tmp);
-		tmp = find_expansion(tmp);
-		if (check_redirections(tmp) == -1)
-		{
-			printf("ganged here");
-			return (-1);
-		}
-		j = recup_delim2(shell, tmp, j);
-	}
-	return (1);
-}
-
 void	file_here(int i, t_here *here)
 {
 	char	*s;
 
-	while(1)
+	while (1)
 	{
 		s = readline("here_doc>");
 		if (!s)
