+++ src/tools/file/free_close.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   free_close.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: zaki <zaki@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: hstephan <hstephan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/15 14:09:02 by zaki              #+#    #+#             */
-/*   Updated: 2023/09/15 18:41:22 by zaki             ###   ########.fr       */
+/*   Updated: 2023/10/02 15:46:28 by hstephan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
