+++ src/tools/file/redirections_utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   redirections_utils.c                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: zaki <zaki@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: hstephan <hstephan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/23 17:22:08 by zlemery           #+#    #+#             */
-/*   Updated: 2023/09/15 17:50:50 by zaki             ###   ########.fr       */
+/*   Updated: 2023/10/02 15:47:09 by hstephan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
